`BukkitElementExtensions`: raw copy-paste
